refactor: rename hyphen-sign experiment to hyphen-digit
