+++ scripts/prompt-experiments/hyphen-digit/2026-09-11/provenance.md
